+++ PortfolioTracker/PortfolioTracker.Logic/Models/RebalanceResult.cs
@@ -4,7 +4,7 @@ namespace PortfolioTracker.Logic.Models
 {
     public class RebalanceResult
     {
-        public decimal RemainingCashOnHand { get; set; }
+        public double RemainingCashOnHand { get; set; }
         public List<RebalanceAction> Actions { get; set; }
     }
 }
