+++ PortfolioTracker/PortfolioTracker.Logic/Models/ScenarioResult.cs
@@ -4,10 +4,10 @@ namespace PortfolioTracker.Logic.Models
 {
     public class ScenarioResult
     {
-        public decimal PercentIncrease { get; set; }
+        public double PercentIncrease { get; set; }
         public DateTimeOffset StartDate { get; set; }
         public DateTimeOffset EndDate { get; set; }
-        public decimal FinalPortfolioValue { get; set; }
-        public decimal TotalCashInvested { get; set; }
+        public double FinalPortfolioValue { get; set; }
+        public double TotalCashInvested { get; set; }
     }
 }
