+++ PortfolioTracker/PortfolioTracker.Models/Portfolio.cs
@@ -3,6 +3,6 @@
     public class Portfolio : Category
     {
         public virtual string Id { get; set; }
-        public decimal CashOnHand { get; set; }
+        public double CashOnHand { get; set; }
     }
 }
